backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.4.0
python-dotenv==0.19.0
openpyxl==3.0.9